+++ src/features/currency/hooks/useQueryParams.ts
@@ -1,8 +1,8 @@
 import { useCallback, useEffect, useMemo } from 'react';
 import { useLocationSearch } from '../../../hooks/useLocationSearch';
-import type { QueryParams } from '../../../types';
 import { parseSearch, buildSearch } from '../lib/search';
 import { usePersistQueryParams } from '../stores/persistQueryParamsStore';
+import type { QueryParams } from '../types';
 
 export type UseQueryParams = {
   queryParams: QueryParams;
