+++ src/hooks/useTitle.ts
@@ -1,7 +1,25 @@
 import { useMemo } from 'react';
 import { useMatches, type UIMatch } from 'react-router';
-import type { RouteHandle } from '../types';
 import { useQueryParams } from '../features/currency/hooks/useQueryParams';
+import type { QueryParams } from '../features/currency/types';
+
+export type RouteHandle = {
+  title?: ((queryParams: QueryParams) => string) | string;
+};
+
+export const createDynamicRouteHandleTitle = (
+  title: string
+): RouteHandle['title'] => {
+  return (queryParams) => {
+    const currencies = queryParams.by.join(', ');
+
+    if (!currencies) {
+      return title;
+    }
+
+    return `${currencies} - ${title}`;
+  };
+};
 
 export const useTitle = (): string => {
   const matches = useMatches() as UIMatch<unknown, RouteHandle>[];
