+++ src/types/statuses.ts
@@ -1,0 +1,1 @@
+export type StatusValue = 'default' | 'failure';
