+++ src/types/themes.ts
@@ -1,0 +1,1 @@
+export type ThemeValue = 'system' | 'light' | 'dark';
